- made sector_t::damageamount an int so that it can hold TELEFRAG_DAMAGE. - marked all places where sector_t::special needs to be addressed for the damage overhaul.
NOTE: This commit will not compile!

+++ src/r_defs.h
@@ -738,7 +738,7 @@ struct sector_t
 
 	float gravity;			// [RH] Sector gravity (1.0 is normal)
 	FNameNoInit damagetype;		// [RH] Means-of-death for applied damage
-	short damageamount;			// [RH] Damage to do while standing on floor
+	int damageamount;			// [RH] Damage to do while standing on floor
 	short damageinterval;	// Interval for damage application
 	short leakydamage;		// chance of leaking through radiation suit
